pyproject.toml
[project]
name = "morpho-net"
version = "0.1.0"
description = "Modular research framework for morphological neural networks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "tensorflow>=2.15",
    "keras>=3.0",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]

[project.scripts]
morpho-train = "morpho_net.main:main"
morpho-evaluate = "morpho_net.evaluate:main"
train = "morpho_net.main:main"
evaluate = "morpho_net.evaluate:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["morpho_net"]
